Fix GUI and Count
