+++ Assets/Scripts/Cam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7a52b8c3e248cd64488470cf05e65768
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
