+++ Assets/Scripts/Projectile.cs
@@ -1,23 +1,15 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using TMPro;
 
 public class Projectile : MonoBehaviour
 {
 
-    private int count;
-    public TextMeshProUGUI countText;
-    public GameObject winTextObject;
+   
 
     public static float bottomY = -13.90522f;                                  // a
 
-    void Start()
-    {
-        count = 0;
-        countText = GameObject.Find("CountText").GetComponent<TextMeshProUGUI>();
-        SetCountText();
-    }
+    
     void Update()
     {
 
@@ -32,20 +24,10 @@ public class Projectile : MonoBehaviour
         if (other.gameObject.CompareTag("PickUp"))
         {
             other.gameObject.SetActive(false);
-            count = count + 1;
-
-            SetCountText();
+            Camera.main.GetComponent<Cam>().SetCountText();
         }
     }
 
-    void SetCountText()
-    {
-        countText.SetText("Count: " + count.ToString());
-
-        if (count >= 10)
-        {
-            winTextObject.SetActive(true);
-        }
-    }
+    
 }
